@@ -27,7 +27,7 @@ project_urls =
     Source Code=https://github.com/Project-MONAI/MONAILabel
 
 [options]
-python_requires = >= 3.8
+python_requires = >= 3.9
 # for compiling and develop setup only
 # no need to specify the versions so that we could
 # compile for multiple targeted versions.
@@ -35,39 +35,43 @@ setup_requires =
     torch
     ninja
 install_requires =
-    monai[nibabel, skimage, pillow, tensorboard, gdown, ignite, torchvision, itk, tqdm, lmdb, psutil, openslide, fire, mlflow]>=1.2.0
-    uvicorn==0.21.1
-    pydantic>=1.10.7
-    python-dotenv==1.0.0
-    fastapi==0.95.0
-    python-multipart==0.0.6
-    requests-toolbelt==0.10.1
-    httpx==0.23.3
-    schedule==1.1.0
-    dicomweb-client==0.59.1
-    timeloop==1.0.2
-    expiringdict==1.2.2
-    expiring_dict==1.1.0
-    cachetools==5.3.0
-    watchdog==3.0.0
-    pydicom==2.3.1
-    pydicom-seg==0.4.1
-    pynetdicom==2.0.2
-    pynrrd==1.0.0
-    opencv-python-headless==4.7.0.72
-    numpymaxflow==0.0.5
-    girder-client==3.1.17
-    ninja==1.11.1
-    einops>=0.6.0
-    pyyaml==6.0
-    filelock==3.11.0
-    passlib==1.7.4
-    python-jose[cryptography]==3.3.0
-    bcrypt==4.0.1
-    shapely==2.0.1
-    requests==2.28.2
+    monai[nibabel, skimage, pillow, tensorboard, gdown, ignite, torchvision, itk, tqdm, lmdb, psutil, openslide, fire, mlflow]>=1.4.0
+    uvicorn>=0.29.0
+    pydantic>=2.7.0
+    pydantic-settings>=2.2.1
+    python-dotenv>=1.0.1
+    fastapi>=0.110.2
+    python-multipart>=0.0.9
+    httpx>=0.27.0
+    schedule>=1.2.1
+    dicomweb-client[gcp]>=0.59.1
+    timeloop>=1.0.2
+    expiringdict>=1.2.2
+    expiring_dict>=1.1.0
+    cachetools>=5.3.3
+    watchdog>=4.0.0
+    pydicom>=2.4.4
+    pydicom-seg>=0.4.1
+    pynetdicom>=2.0.2
+    pynrrd>=1.0.0
+    numpymaxflow>=0.0.7
+    girder-client>=3.2.3
+    ninja>=1.11.1.1
+    einops>=0.7.0
+    pyyaml>=6.0.1
+    filelock>=3.11.0
+    passlib>=1.7.4
+    pyjwt>=2.8.0
+    bcrypt>=4.1.2
+    shapely>=2.0.4
+    requests>=2.31.0
+    requests-toolbelt>=1.0.0
+    urllib3>=2.2.1
     scikit-learn
     scipy
+    google-auth>=2.29.0
+    sam2>=0.4.1; python_version >= '3.10'
+    #SAM-2 @ git+https://github.com/facebookresearch/sam2.git@c2ec8e14a185632b0a5d8b161928ceb50197eddc ; python_version >= '3.10'
 
 [flake8]
 select = B,C,E,F,N,P,T4,W,B9
