@@ -22,6 +22,7 @@ from xml.etree import ElementTree
 import requests
 from requests.auth import HTTPBasicAuth
 
+from monailabel.datastore.utils.convert import nifti_to_dicom_seg
 from monailabel.interfaces.datastore import Datastore
 from monailabel.utils.others.generic import md5_digest
 
@@ -158,8 +159,33 @@ class XNATDatastore(Datastore):
     def remove_image(self, image_id: str) -> None:
         raise NotImplementedError
 
+    def __convert_nifti_to_dcmseg(self, series_dir, nii_seg_path, model_name, label_names) -> str:
+        label_info = []
+        for i, lb in enumerate(label_names):
+            label_info.append(
+                {"model_name": model_name, "name": str(i + 1) + "_" + lb, "description": "lb" + str(i + 1) + "_" + lb}
+            )
+        dcmSegFile = nifti_to_dicom_seg(series_dir=series_dir, label=nii_seg_path, label_info=label_info)
+        logging.info(f" converted nifti to dicom seg --- at {dcmSegFile}")
+        return dcmSegFile
+
     def save_label(self, image_id: str, label_filename: str, label_tag: str, label_info: Dict[str, Any]) -> str:
-        raise NotImplementedError
+        aiaa_model_name = label_info.get("model", "NoModel")
+        label_names = label_info.get("params", {}).get("label_names", {})
+
+        # save the nii.gz segmentation into Xnat
+        project, subject, experiment, scan = self._id_to_fields(image_id)
+        nameAtXnat = f"pat_{subject}_exp_{experiment}_S_{scan}_AI_{aiaa_model_name}.nii.gz"
+        self._request_put_file(
+            experiment, scan, name_at_xnat=nameAtXnat, file2send=label_filename, ai_model_name=aiaa_model_name
+        )
+        # convert nii to dcm seg and upload to Xnat
+        if label_filename.endswith(".nii") or label_filename.endswith(".nii.gz"):
+            series_dir = self._download_image(image_id)
+            tmp_dcm_segpath = self.__convert_nifti_to_dcmseg(series_dir, label_filename, aiaa_model_name, label_names)
+            self.__upload_assessment(aiaa_model_name, image_id, tmp_dcm_segpath, "SEG")
+
+        return image_id
 
     def remove_label(self, label_id: str, label_tag: str) -> None:
         raise NotImplementedError
@@ -292,10 +318,73 @@ class XNATDatastore(Datastore):
         logger.info("XNAT:: Logged in XNAT")
 
     def _request_get(self, url):
-        return self.xnat_session.get(url, allow_redirects=True)
+        return self.xnat_session.get(url, auth=self.auth, allow_redirects=True)
 
     def _request_post(self, url):
         return self.xnat_session.post(url, auth=self.auth, allow_redirects=True)
+
+    def _request_put(self, url, data, type):
+        response = self.xnat_session.put(
+            url,
+            data=data,
+            params={"overwrite": "true", "type": type},
+            headers={"Content-Type": "application/octet-stream"},
+            auth=self.auth,
+            allow_redirects=True,
+        )
+        if response.status_code != 200:  # failed call
+            logger.error(f" xnat put call error status_code= {response.status_code}  text ={response.text}")
+        else:
+            logger.info(f" xnat dcm-seg / measurement json put completed {response.text}")
+        return response
+
+    def _request_put_file(self, experiment, scan, file2send, name_at_xnat, ai_model_name):
+        """
+        uploads file to xnat using REST API
+        """
+        folder = "AI"
+        url = "{}/REST/experiments/{}/scans/{}/resources/{}/files/{}".format(
+            self.api_url,
+            quote_plus(experiment),
+            quote_plus(scan),
+            quote_plus(folder),
+            quote_plus(name_at_xnat),
+        )
+        data = open(file2send, "rb")
+        params = {"overwrite": "true", "description": name_at_xnat, "content": ai_model_name, "format": "nii"}
+        response = self.xnat_session.put(
+            url, params=params, data=data, headers={"Content-Type": "application/octet-stream"}, allow_redirects=True
+        )
+        if response.status_code != 200:  # failed call
+            logger.error(f" put call error status_code= {response.status_code}  text ={response.text}")
+        else:
+            logger.info(f" put completed {response.text}")
+
+        return response
+
+    def __upload_assessment(self, aiaa_model_name, image_id, file_path, type):
+        """
+        uploads Assessments to xnat using xapi
+        :param aiaa_model_name: model name used to generate this file
+        :param image_id: image id to use
+        :param file_path: file to upload
+        :param type: "SEG" for dicom seg or "MEAS" for measurments
+        """
+        if not os.path.exists(file_path):
+            logging.error(f" file {file_path} does not exist! ")
+            return
+        data = open(file_path, "rb")
+        project, subject, experiment, scan = self._id_to_fields(image_id)
+        url = "{}/xapi/roi/projects/{}/sessions/{}/collections/Pat{}_S{}_{}".format(
+            self.api_url,
+            quote_plus(project),
+            quote_plus(experiment),
+            quote_plus(subject),
+            quote_plus(scan),
+            quote_plus(aiaa_model_name),
+        )
+
+        self._request_put(url, data, type=type)
 
 
 """
