@@ -16,13 +16,12 @@ import numpy as np
 import torch
 from monai.apps.deepgrow.transforms import AddRandomGuidanced, FindDiscrepancyRegionsd
 from monai.handlers import MeanDice, from_engine
-from monai.handlers.ignite_metric import IgniteMetric
+from monai.handlers.ignite_metric import IgniteMetricHandler
 from monai.inferers import SimpleInferer
 from monai.losses import DiceLoss
 from monai.metrics import MeanIoU
 from monai.transforms import (
     Activationsd,
-    AddChanneld,
     AsDiscreted,
     EnsureChannelFirstd,
     EnsureTyped,
@@ -88,9 +87,9 @@ class DeepEdit(BasicTrainTask):
 
     def train_pre_transforms(self, context: Context):
         return [
-            LoadImaged(keys=("image", "label"), dtype=np.uint8),
+            LoadImaged(keys=("image", "label"), dtype=np.uint8, image_only=False),
             EnsureChannelFirstd(keys="image"),
-            AddChanneld(keys="label"),
+            EnsureChannelFirstd(keys="label", channel_dim="no_channel"),
             Resized(keys=("image", "label"), spatial_size=self.roi_size, mode=("area", "nearest")),
             ToTensord(keys="image"),
             TorchVisiond(
@@ -150,7 +149,7 @@ class DeepEdit(BasicTrainTask):
         )
 
 
-class MeanIoUMetric(IgniteMetric):
+class MeanIoUMetric(IgniteMetricHandler):
     def __init__(
         self,
         include_background: bool = True,
