@@ -9,13 +9,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import csv
+import json
 import logging
 import os
 from datetime import timedelta
-from typing import Dict
+from typing import Dict, List, Union
 
 import lib.configs
+import numpy as np
 import schedule
+from monai.transforms import ToNumpyd
 from timeloop import Timeloop
 
 import monailabel
@@ -24,10 +28,11 @@ from monailabel.datastore.cvat import CVATDatastore
 from monailabel.interfaces.app import MONAILabelApp
 from monailabel.interfaces.config import TaskConfig
 from monailabel.interfaces.datastore import Datastore
-from monailabel.interfaces.tasks.infer_v2 import InferTask
+from monailabel.interfaces.tasks.infer_v2 import InferTask, InferType
 from monailabel.interfaces.tasks.scoring import ScoringMethod
 from monailabel.interfaces.tasks.strategy import Strategy
 from monailabel.interfaces.tasks.train import TrainTask
+from monailabel.sam2.utils import is_sam2_module_available
 from monailabel.tasks.activelearning.random import Random
 from monailabel.utils.others.class_utils import get_class_names
 from monailabel.utils.others.generic import create_dataset_from_path, strtobool
@@ -55,9 +60,9 @@ class MyApp(MONAILabelApp):
             print(f"    all, {', '.join(configs.keys())}")
             print("---------------------------------------------------------------------------------------")
             print("")
-            exit(-1)
+            # exit(-1)
 
-        models = models.split(",")
+        models = models.split(",") if models else []
         models = [m.strip() for m in models]
         invalid = [m for m in models if m != "all" and not configs.get(m)]
         if invalid:
@@ -82,6 +87,7 @@ class MyApp(MONAILabelApp):
 
         logger.info(f"+++ Using Models: {list(self.models.keys())}")
 
+        self.sam = strtobool(conf.get("sam2", "true"))
         super().__init__(
             app_dir=app_dir,
             studies=studies,
@@ -92,9 +98,64 @@ class MyApp(MONAILabelApp):
         )
         self.downloading = False
 
+    def read_labels_from_file(self, file_path: str) -> List[Dict[str, Union[str, int, list]]]:
+        labels = []
+
+        try:
+            with open(file_path) as csvfile:
+                reader = csv.DictReader(csvfile)
+
+                for row in reader:
+                    # Check for required fields
+                    if "name" in row and "id" in row and "color" in row:
+                        attributes = []
+                        # Process the attributes if they exist and are not empty
+                        if "attributes" in row and row["attributes"].strip():
+                            try:
+                                attributes = json.loads(row["attributes"].strip())
+                            except json.JSONDecodeError as e:
+                                logger.warning(f"Could not parse attributes JSON for row {row}: {e}")
+
+                        entry = {
+                            "name": row["name"],
+                            "id": int(row["id"]),
+                            "color": row["color"],
+                            "type": "any",
+                            "attributes": attributes,
+                        }
+                        labels.append(entry)
+                    else:
+                        logger.warning(f"Skipping row due to missing fields: {row}")
+
+                logger.info(f"Loaded {len(labels)} labels from {file_path}")
+        except FileNotFoundError:
+            logger.error(f"Label file {file_path} not found!")
+        except Exception as e:
+            logger.error(f"Error reading label file {file_path}: {e}")
+
+        return labels
+
     def init_datastore(self) -> Datastore:
         if settings.MONAI_LABEL_DATASTORE_URL and settings.MONAI_LABEL_DATASTORE.lower() == "cvat":
             logger.info(f"Using CVAT: {self.studies}")
+
+            labels: List[Dict[str, Union[str, int, list]]] = []
+
+            # If cvat_labels_file is specified and not null, read labels from that file
+            cvat_labels_file = self.conf.get("cvat_labels_file")
+            if cvat_labels_file:
+                try:
+                    labels = self.read_labels_from_file(cvat_labels_file)
+                except Exception as e:
+                    logger.error(f"Error reading CVAT labels file {cvat_labels_file}: {e}")
+            else:
+                label_str = self.conf.get("cvat_labels")
+                if label_str:
+                    try:
+                        labels = json.loads(label_str)
+                    except Exception as e:
+                        logger.error(f"Error parsing cvat_labels config: {e}")
+
             return CVATDatastore(
                 datastore_path=self.studies,
                 api_url=settings.MONAI_LABEL_DATASTORE_URL,
@@ -103,7 +164,7 @@ class MyApp(MONAILabelApp):
                 project=self.conf.get("cvat_project", "MONAILabel"),
                 task_prefix=self.conf.get("cvat_task_prefix", "ActiveLearning_Iteration"),
                 image_quality=int(self.conf.get("cvat_image_quality", "70")),
-                labels=self.conf.get("cvat_labels"),
+                labels=labels,
                 normalize_label=strtobool(self.conf.get("cvat_normalize_label", "true")),
                 segment_size=int(self.conf.get("cvat_segment_size", "1")),
                 extensions=settings.MONAI_LABEL_DATASTORE_FILE_EXT,
@@ -123,6 +184,21 @@ class MyApp(MONAILabelApp):
             for k, v in c.items():
                 logger.info(f"+++ Adding Inferer:: {k} => {v}")
                 infers[k] = v
+
+        #################################################
+        # SAM
+        #################################################
+        if is_sam2_module_available() and self.sam:
+            from monailabel.sam2.infer import Sam2InferTask
+
+            infers["sam_2d"] = Sam2InferTask(
+                model_dir=self.model_dir,
+                type=InferType.ANNOTATION,
+                dimension=2,
+                post_trans=[ToNumpyd(keys="pred", dtype=np.uint8)],
+                config={"cache_image": False, "reset_state": True},
+            )
+
         return infers
 
     def init_trainers(self) -> Dict[str, TrainTask]:
