@@ -31,7 +31,7 @@ class NuClick(TaskConfig):
 
         bundle_name = "pathology_nuclick_annotation"
         zoo_source = conf.get("zoo_source", settings.MONAI_ZOO_SOURCE)
-        version = conf.get("nuclick", "0.1.1")
+        version = conf.get("nuclick", "0.2.1")
 
         self.bundle_path = os.path.join(self.model_dir, bundle_name)
         if not os.path.exists(self.bundle_path):
