@@ -21,7 +21,7 @@ from monai.utils import optional_import
 from monailabel.interfaces.config import TaskConfig
 from monailabel.interfaces.tasks.infer_v2 import InferTask
 from monailabel.interfaces.tasks.train import TrainTask
-from monailabel.utils.others.generic import download_file, strtobool
+from monailabel.utils.others.generic import download_file, remove_file, strtobool
 
 _, has_cp = optional_import("cupy")
 _, has_cucim = optional_import("cucim")
@@ -34,33 +34,38 @@ class Segmentation(TaskConfig):
         super().init(name, model_dir, conf, planner, **kwargs)
 
         # Labels
-        self.labels = {
-            "spleen": 1,
-            "kidney_right": 2,
-            "kidney_left": 3,
-            "gallbladder": 4,
-            "liver": 5,
-            "stomach": 6,
-            "aorta": 7,
-            "inferior_vena_cava": 8,
-            "portal_vein_and_splenic_vein": 9,
-            "pancreas": 10,
-            "adrenal_gland_right": 11,
-            "adrenal_gland_left": 12,
-            "lung_upper_lobe_left": 13,
-            "lung_lower_lobe_left": 14,
-            "lung_upper_lobe_right": 15,
-            "lung_middle_lobe_right": 16,
-            "lung_lower_lobe_right": 17,
-            "esophagus": 42,
-            "trachea": 43,
-            "heart_myocardium": 44,
-            "heart_atrium_left": 45,
-            "heart_ventricle_left": 46,
-            "heart_atrium_right": 47,
-            "heart_ventricle_right": 48,
-            "pulmonary_artery": 49,
-        }
+        conf_labels = self.conf.get("labels")
+        self.labels = (
+            {label: idx for idx, label in enumerate(conf_labels.split(","), start=1)}
+            if conf_labels
+            else {
+                "spleen": 1,
+                "kidney_right": 2,
+                "kidney_left": 3,
+                "gallbladder": 4,
+                "liver": 5,
+                "stomach": 6,
+                "aorta": 7,
+                "inferior_vena_cava": 8,
+                "portal_vein_and_splenic_vein": 9,
+                "pancreas": 10,
+                "adrenal_gland_right": 11,
+                "adrenal_gland_left": 12,
+                "lung_upper_lobe_left": 13,
+                "lung_lower_lobe_left": 14,
+                "lung_upper_lobe_right": 15,
+                "lung_middle_lobe_right": 16,
+                "lung_lower_lobe_right": 17,
+                "esophagus": 42,
+                "trachea": 43,
+                "heart_myocardium": 44,
+                "heart_atrium_left": 45,
+                "heart_ventricle_left": 46,
+                "heart_atrium_right": 47,
+                "heart_ventricle_right": 48,
+                "pulmonary_artery": 49,
+            }
+        )
 
         # Model Files
         self.path = [
@@ -69,10 +74,14 @@ class Segmentation(TaskConfig):
         ]
 
         # Download PreTrained Model
-        if strtobool(self.conf.get("use_pretrained_model", "true")):
+        if not conf_labels and strtobool(self.conf.get("use_pretrained_model", "true")):
             url = f"{self.conf.get('pretrained_path', self.PRE_TRAINED_PATH)}"
             url = f"{url}/radiology_segmentation_segresnet_multilabel.pt"
             download_file(url, self.path[0])
+
+        # Remove pre-trained pt if user is using his/her custom labels.
+        if conf_labels:
+            remove_file(self.path[0])
 
         self.target_spacing = (1.5, 1.5, 1.5)  # target space for image
         # Setting ROI size - This is for the image padding
