@@ -11,6 +11,7 @@
 
 from typing import Callable, Sequence
 
+from lib.transforms.transforms import GetCentroidsd
 from monai.inferers import Inferer, SlidingWindowInferer
 from monai.transforms import (
     Activationsd,
@@ -93,11 +94,14 @@ class Segmentation(BasicInferTask):
 
         if data and data.get("largest_cc", False):
             t.append(KeepLargestConnectedComponentd(keys="pred"))
-        t.append(
-            Restored(
-                keys="pred",
-                ref_image="image",
-                config_labels=self.labels if data.get("restore_label_idx", False) else None,
-            )
+        t.extend(
+            [
+                Restored(
+                    keys="pred",
+                    ref_image="image",
+                    config_labels=self.labels if data.get("restore_label_idx", False) else None,
+                ),
+                GetCentroidsd(keys="pred", centroids_key="centroids"),
+            ]
         )
         return t
