@@ -34,8 +34,8 @@ class TestConvert(unittest.TestCase):
 
     def test_binary_to_image(self):
         reference_image = os.path.join(self.local_dataset, "labels", "final", "spleen_3.nii.gz")
-        label, _ = LoadImage()(reference_image)
-        label = label.astype(np.uint16)
+        label = LoadImage(image_only=True)(reference_image)
+        label = label.astype(np.uint8)
         label = label.flatten(order="F")
 
         label_bin = tempfile.NamedTemporaryFile(suffix=".bin").name
